+++ libs/sabu_chassis/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.1.1] - 2026-05-26
+
+### Added
+- Added `get_logger()` for library-safe logging with sabu-chassis formatting
+  when an application configures logging.
+
 ## [0.1.0] - 2026-04-10
 
 ### Added
@@ -10,6 +16,5 @@
 - This package is a continuation and rename of `jugs-chassis`
 - For full history prior to this release, see:
   https://pypi.org/project/jugs-chassis/
-